+++ src/py42/services/users.py
@@ -2,6 +2,7 @@ from py42 import settings
 from py42._compat import quote
 from py42.services import BaseService
 from py42.services.util import get_all_pages
+from py42.exceptions import Py42UserAlreadyExistsError
 
 
 class UserService(BaseService):
@@ -47,7 +48,12 @@ class UserService(BaseService):
             u"lastName": last_name,
             u"notes": notes,
         }
-        return self._connection.post(uri, json=data)
+
+        response = self._connection.post(uri, json=data)
+        if response.status_code == 500 and response.text == "USER_DUPLICATE":
+            raise Py42UserAlreadyExistsError(response)
+
+        return response
 
     def get_by_id(self, user_id, **kwargs):
         """Gets the user with the given ID.